[B2B-API] update on api added on user if what type of roles they have and passed onto the json response and hidden other data [B2B-API] added new API to delete product image and on the DB

+++ app/Repositories/ProductRepository.php
@@ -440,11 +440,15 @@ class ProductRepository implements ProductInterface
         DB::beginTransaction();
         try {
             $user = Auth::user();
+
             $product = Product::where('id',$id)->where('user_id',$user->id)->first();
 
             if (!$product) return $this->error("Product not found", 404);
 
             app(PriceRepository::class)->deletePrice($id);
+
+            $pimg = app(ProductImageRepository::class)->deleteByProduct($product);
+            if($pimg) return $pimg;
 
             $product->delete();
